Refactoring MarkovBuilder as a public-interface wrapper
MarkovBuilder instances now wraps an instance of MarkovProcess, allowing what is 
intended to be thread safe public access to the modification functions.

All publically available instances of MarkovProcess should be immuatable.

The MarkovBuilder only returns the underlying MarkovProcess directly after MarkovBuilder.seal() has been called.
seal() sets the sealed flag, which prevents any further alterations to the MarkovProceses structure.

Before or after seal() is called, a MarkovProcess with the same structure can be obtained with getProcess().
Before seal(), this uses the clone() function of MarkovProcess. After seal(), this returns a reference to the builders
internal MarkovProcess.